+++ lib/pangea/analytics_misc/construct_use_type_enum.dart
@@ -508,21 +508,6 @@ extension ConstructUseTypeExtension on ConstructUseTypeEnum {
   }
 }
 
-/// Priority tier for spaced repetition scoring.
-///
-/// Tier is determined by how the user most recently encountered the word
-/// in chat (wa/ga/ta) and whether they've struggled with it in practice.
-enum PracticeTier {
-  /// User demonstrated mastery (wa) with no subsequent errors — skip entirely.
-  suppressed,
-
-  /// User needed help (ta/ga) or recently got it wrong — prioritize.
-  active,
-
-  /// Standard aging-based priority (correctly practiced, but aging).
-  maintenance,
-}
-
 class ConstructUseTypeUtil {
   static ConstructUseTypeEnum fromString(String value) {
     return ConstructUseTypeEnum.values.firstWhere(
